+++ Common/src/main/java/ca/lukegrahamlandry/lib/data/impl/TileFileDataWrapper.java
@@ -1,5 +1,15 @@
+/*
+ * This file is part of WrapperLib
+ * Copyright 2022 LukeGrahamLandry
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
 package ca.lukegrahamlandry.lib.data.impl;
 
+import ca.lukegrahamlandry.lib.base.WorkInProgress;
 import ca.lukegrahamlandry.lib.data.sync.TileFileDataMessage;
 import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceLocation;
@@ -7,6 +17,7 @@ import net.minecraft.world.level.block.entity.BlockEntity;
 
 import java.util.Map;
 
+@WorkInProgress
 public class TileFileDataWrapper<V> extends LevelDataWrapper<Map<BlockPos, V>> {
     public TileFileDataWrapper(Class<Map<BlockPos, V>> clazz) {
         super(clazz);